Rename credentials POST handler to reflect targetId path param
