+++ src/main/java/com/redhat/rhjmc/containerjfr/net/web/http/api/v2/HttpApiV2Module.java
@@ -101,15 +101,15 @@ public abstract class HttpApiV2Module {
 
     @Binds
     @IntoSet
-    abstract RequestHandler bindCredentialsPostHandler(CredentialsPostHandler handler);
+    abstract RequestHandler bindTargetCredentialsPostHandler(TargetCredentialsPostHandler handler);
 
     @Binds
     @IntoSet
-    abstract RequestHandler bindCredentialsPostBodyHandler(CredentialsPostBodyHandler handler);
+    abstract RequestHandler bindTargetCredentialsPostBodyHandler(TargetCredentialsPostBodyHandler handler);
 
     @Binds
     @IntoSet
-    abstract RequestHandler bindCredentialsDeleteHandler(CredentialsDeleteHandler handler);
+    abstract RequestHandler bindTargetCredentialsDeleteHandler(TargetCredentialsDeleteHandler handler);
 
     @Binds
     @IntoSet
