+++ src/main/java/com/redhat/rhjmc/containerjfr/net/web/http/api/v2/TargetCredentialsPostBodyHandler.java
@@ -41,40 +41,28 @@
  */
 package com.redhat.rhjmc.containerjfr.net.web.http.api.v2;
 
-import java.io.IOException;
-import java.util.Objects;
-
 import javax.inject.Inject;
 
-import org.apache.commons.lang3.StringUtils;
-import com.google.gson.Gson;
-
-import com.redhat.rhjmc.containerjfr.configuration.CredentialsManager;
-import com.redhat.rhjmc.containerjfr.core.log.Logger;
-import com.redhat.rhjmc.containerjfr.core.net.Credentials;
 import com.redhat.rhjmc.containerjfr.net.AuthManager;
-import com.redhat.rhjmc.containerjfr.net.web.http.HttpMimeType;
+import com.redhat.rhjmc.containerjfr.net.web.http.AbstractAuthenticatedRequestHandler;
 import com.redhat.rhjmc.containerjfr.net.web.http.api.ApiVersion;
+
 import io.vertx.core.http.HttpMethod;
+import io.vertx.ext.web.RoutingContext;
+import io.vertx.ext.web.handler.BodyHandler;
 
-class CredentialsPostHandler extends AbstractV2RequestHandler<Void> {
+class TargetCredentialsPostBodyHandler extends AbstractAuthenticatedRequestHandler {
 
-    static final String PATH = "targets/:targetId/credentials";
-
-    private final CredentialsManager credentialsManager;
-    private final Logger logger;
+    static final BodyHandler BODY_HANDLER = BodyHandler.create(true);
 
     @Inject
-    CredentialsPostHandler(
-            AuthManager auth, CredentialsManager credentialsManager, Gson gson, Logger logger) {
-        super(auth, gson);
-        this.credentialsManager = credentialsManager;
-        this.logger = logger;
+    TargetCredentialsPostBodyHandler(AuthManager auth) {
+        super(auth);
     }
 
     @Override
-    public boolean requiresAuthentication() {
-        return true;
+    public int getPriority() {
+        return DEFAULT_PRIORITY - 1;
     }
 
     @Override
@@ -89,51 +77,11 @@ class CredentialsPostHandler extends AbstractV2RequestHandler<Void> {
 
     @Override
     public String path() {
-        return basePath() + PATH;
+        return basePath() + TargetCredentialsPostHandler.PATH;
     }
 
     @Override
-    public HttpMimeType mimeType() {
-        return HttpMimeType.PLAINTEXT;
-    }
-
-    @Override
-    public boolean isAsync() {
-        return false;
-    }
-
-    @Override
-    public boolean isOrdered() {
-        return true;
-    }
-
-    @Override
-    public IntermediateResponse<Void> handle(RequestParameters params) throws ApiException {
-        String targetId = params.getPathParams().get("targetId");
-        String username;
-        String password;
-        try {
-            username =
-                    Objects.requireNonNull(
-                            params.getFormAttributes().get("username"), "Username is required");
-            password =
-                    Objects.requireNonNull(
-                            params.getFormAttributes().get("password"), "Password is required");
-        } catch (NullPointerException npe) {
-            throw new ApiException(400, npe.getMessage(), npe);
-        }
-        boolean persist =
-                Boolean.valueOf(
-                        StringUtils.defaultString(
-                                params.getFormAttributes().get("persist"), "false"));
-
-        try {
-            this.credentialsManager.addCredentials(
-                    targetId, new Credentials(username, password), persist);
-        } catch (IOException e) {
-            throw new ApiException(500, "IOException occurred while persisting credentials", e);
-        }
-
-        return new IntermediateResponse<Void>().body(null);
+    public void handleAuthenticated(RoutingContext ctx) throws Exception {
+        BODY_HANDLER.handle(ctx);
     }
 }
